move ext module into its own crate
