+++ adapter/ph/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 arrayref = { version = "0.3.7" }
 bytes = { version = "1.7.1" }
+cbpf-rs = { path = "../../cbpf-rs", features = ["pcap"] }
 clap = { version = "4.5.7", features = ["derive"] }
 enum-map = { version = "2.7.3" }
 foreign-types = { version = "0.3.2" }
@@ -21,7 +22,7 @@ tokio-tun = { version = "0.11.5" }
 tokio = { version = "1.37.0", features = ["full"] }
 zerocopy-derive = { version = "0.7.34" }
 zerocopy = { version = "0.7.34", features = ["byteorder"] }
-cbpf-rs = { path = "../../cbpf-rs", features = ['pcap'] }
+zpr-ext = { path = "../../zpr-ext", features = ["bytes", "tokio", "tokio-tun", "zerocopy"] }
 
 [features]
 ci = []
