+++ adapter/ph/src/packet.rs
@@ -1,12 +1,12 @@
 //! This module contains all state of a packet which is moving through the system.
 
 use crate::config;
-use crate::ext::std::mem::DropGuard;
 use crate::net_defs::*;
 use bytes::buf;
 use std::mem::{size_of, size_of_val};
 use zerocopy::{AsBytes, ByteOrder, FromBytes, FromZeroes, NetworkEndian};
 use zerocopy_derive::{AsBytes, FromBytes, FromZeroes};
+use zpr_ext::std::mem::DropGuard;
 
 /** Exclusive handle to an in-use packet buffer.
  *
