+++ zpr-ext/src/lib.rs
@@ -1,5 +1,11 @@
 /// Various "extensions" to external crates.
 pub mod std;
+
+#[cfg(feature = "tokio")]
 pub mod tokio;
+
+#[cfg(feature = "tokio-tun")]
 pub mod tokio_tun;
+
+#[cfg(feature = "zerocopy")]
 pub mod zerocopy;
